+++ packages/iov-jsonrpc/types/types.d.ts
@@ -1,13 +1,14 @@
 import { JsonCompatibleArray, JsonCompatibleDictionary, JsonCompatibleValue } from "./jsoncompatibledictionary";
+export declare type JsonRpcId = number | string;
 export interface JsonRpcRequest {
     readonly jsonrpc: "2.0";
-    readonly id: number;
+    readonly id: JsonRpcId;
     readonly method: string;
     readonly params: JsonCompatibleArray | JsonCompatibleDictionary;
 }
 export interface JsonRpcSuccessResponse {
     readonly jsonrpc: "2.0";
-    readonly id: number;
+    readonly id: JsonRpcId;
     readonly result: any;
 }
 export interface JsonRpcError {
@@ -20,7 +21,7 @@ export interface JsonRpcError {
  */
 export interface JsonRpcErrorResponse {
     readonly jsonrpc: "2.0";
-    readonly id: number | null;
+    readonly id: JsonRpcId | null;
     readonly error: JsonRpcError;
 }
 export declare type JsonRpcResponse = JsonRpcSuccessResponse | JsonRpcErrorResponse;