+++ packages/iov-tendermint-rpc/src/adaptor.ts
@@ -1,9 +1,11 @@
 // This module exposes translators for multiple tendermint versions
 // Pick a version that matches the server to properly encode the data types
 
-import { JsonRpcEvent, JsonRpcRequest, JsonRpcSuccess } from "./jsonrpc";
+import { JsonRpcRequest, JsonRpcSuccessResponse } from "@iov/jsonrpc";
+
 import * as requests from "./requests";
 import * as responses from "./responses";
+import { SubscriptionEvent } from "./rpcclients";
 import { TxBytes, TxHash } from "./types";
 import { v0_25 } from "./v0-25";
 import { v0_27 } from "./v0-27";
@@ -18,7 +20,7 @@ export interface Adaptor {
 export type Encoder<T extends requests.Request> = (req: T) => JsonRpcRequest;
 
 // Decoder is a generic that matches all methods of Responses
-export type Decoder<T extends responses.Response> = (res: JsonRpcSuccess) => T;
+export type Decoder<T extends responses.Response> = (res: JsonRpcSuccessResponse) => T;
 
 export interface Params {
   readonly encodeAbciInfo: (req: requests.AbciInfoRequest) => JsonRpcRequest;
@@ -38,26 +40,26 @@ export interface Params {
 }
 
 export interface Responses {
-  readonly decodeAbciInfo: (response: JsonRpcSuccess) => responses.AbciInfoResponse;
-  readonly decodeAbciQuery: (response: JsonRpcSuccess) => responses.AbciQueryResponse;
-  readonly decodeBlock: (response: JsonRpcSuccess) => responses.BlockResponse;
-  readonly decodeBlockResults: (response: JsonRpcSuccess) => responses.BlockResultsResponse;
-  readonly decodeBlockchain: (response: JsonRpcSuccess) => responses.BlockchainResponse;
-  readonly decodeBroadcastTxSync: (response: JsonRpcSuccess) => responses.BroadcastTxSyncResponse;
-  readonly decodeBroadcastTxAsync: (response: JsonRpcSuccess) => responses.BroadcastTxAsyncResponse;
-  readonly decodeBroadcastTxCommit: (response: JsonRpcSuccess) => responses.BroadcastTxCommitResponse;
-  readonly decodeCommit: (response: JsonRpcSuccess) => responses.CommitResponse;
-  readonly decodeGenesis: (response: JsonRpcSuccess) => responses.GenesisResponse;
-  readonly decodeHealth: (response: JsonRpcSuccess) => responses.HealthResponse;
-  readonly decodeStatus: (response: JsonRpcSuccess) => responses.StatusResponse;
-  readonly decodeTx: (response: JsonRpcSuccess) => responses.TxResponse;
-  readonly decodeTxSearch: (response: JsonRpcSuccess) => responses.TxSearchResponse;
-  readonly decodeValidators: (response: JsonRpcSuccess) => responses.ValidatorsResponse;
+  readonly decodeAbciInfo: (response: JsonRpcSuccessResponse) => responses.AbciInfoResponse;
+  readonly decodeAbciQuery: (response: JsonRpcSuccessResponse) => responses.AbciQueryResponse;
+  readonly decodeBlock: (response: JsonRpcSuccessResponse) => responses.BlockResponse;
+  readonly decodeBlockResults: (response: JsonRpcSuccessResponse) => responses.BlockResultsResponse;
+  readonly decodeBlockchain: (response: JsonRpcSuccessResponse) => responses.BlockchainResponse;
+  readonly decodeBroadcastTxSync: (response: JsonRpcSuccessResponse) => responses.BroadcastTxSyncResponse;
+  readonly decodeBroadcastTxAsync: (response: JsonRpcSuccessResponse) => responses.BroadcastTxAsyncResponse;
+  readonly decodeBroadcastTxCommit: (response: JsonRpcSuccessResponse) => responses.BroadcastTxCommitResponse;
+  readonly decodeCommit: (response: JsonRpcSuccessResponse) => responses.CommitResponse;
+  readonly decodeGenesis: (response: JsonRpcSuccessResponse) => responses.GenesisResponse;
+  readonly decodeHealth: (response: JsonRpcSuccessResponse) => responses.HealthResponse;
+  readonly decodeStatus: (response: JsonRpcSuccessResponse) => responses.StatusResponse;
+  readonly decodeTx: (response: JsonRpcSuccessResponse) => responses.TxResponse;
+  readonly decodeTxSearch: (response: JsonRpcSuccessResponse) => responses.TxSearchResponse;
+  readonly decodeValidators: (response: JsonRpcSuccessResponse) => responses.ValidatorsResponse;
 
   // events
-  readonly decodeNewBlockEvent: (response: JsonRpcEvent) => responses.NewBlockEvent;
-  readonly decodeNewBlockHeaderEvent: (response: JsonRpcEvent) => responses.NewBlockHeaderEvent;
-  readonly decodeTxEvent: (response: JsonRpcEvent) => responses.TxEvent;
+  readonly decodeNewBlockEvent: (response: SubscriptionEvent) => responses.NewBlockEvent;
+  readonly decodeNewBlockHeaderEvent: (response: SubscriptionEvent) => responses.NewBlockHeaderEvent;
+  readonly decodeTxEvent: (response: SubscriptionEvent) => responses.TxEvent;
 }
 
 /**