+++ packages/iov-tendermint-rpc/src/rpcclients/httpclient.spec.ts
@@ -1,5 +1,5 @@
 import { defaultInstance } from "../config.spec";
-import { jsonRpcWith } from "../jsonrpc";
+import { createJsonRpcRequest } from "../jsonrpc";
 import { Method } from "../requests";
 
 import { HttpClient } from "./httpclient";
@@ -21,15 +21,15 @@ describe("HttpClient", () => {
     pendingWithoutTendermint();
     const client = new HttpClient(tendermintUrl);
 
-    const healthResponse = await client.execute(jsonRpcWith(Method.Health));
+    const healthResponse = await client.execute(createJsonRpcRequest(Method.Health));
     expect(healthResponse.result).toEqual({});
 
-    const statusResponse = await client.execute(jsonRpcWith(Method.Status));
+    const statusResponse = await client.execute(createJsonRpcRequest(Method.Status));
     expect(statusResponse.result).toBeTruthy();
     expect(statusResponse.result.node_info).toBeTruthy();
 
     await client
-      .execute(jsonRpcWith("no-such-method"))
+      .execute(createJsonRpcRequest("no-such-method"))
       .then(() => fail("must not resolve"))
       .catch(error => expect(error).toBeTruthy());