+++ packages/iov-tendermint-rpc/src/v0-25/requests.ts
@@ -1,69 +1,70 @@
 import { Encoding } from "@iov/encoding";
+import { JsonRpcRequest } from "@iov/jsonrpc";
 
 import { assertNotEmpty, Base64, Base64String, HexString, Integer, IntegerString, may } from "../encodings";
-import { JsonRpcRequest, jsonRpcWith } from "../jsonrpc";
+import { createJsonRpcRequest } from "../jsonrpc";
 import * as requests from "../requests";
 
 /***** queries *****/
 
 export class Params {
   public static encodeAbciInfo(req: requests.AbciInfoRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method);
+    return createJsonRpcRequest(req.method);
   }
 
   public static encodeAbciQuery(req: requests.AbciQueryRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeAbciQueryParams(req.params));
+    return createJsonRpcRequest(req.method, encodeAbciQueryParams(req.params));
   }
 
   public static encodeBlock(req: requests.BlockRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeHeightParam(req.params));
+    return createJsonRpcRequest(req.method, encodeHeightParam(req.params));
   }
 
   public static encodeBlockchain(req: requests.BlockchainRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeBlockchainRequestParams(req.params));
+    return createJsonRpcRequest(req.method, encodeBlockchainRequestParams(req.params));
   }
 
   public static encodeBlockResults(req: requests.BlockResultsRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeHeightParam(req.params));
+    return createJsonRpcRequest(req.method, encodeHeightParam(req.params));
   }
 
   public static encodeBroadcastTx(req: requests.BroadcastTxRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeBroadcastTxParams(req.params));
+    return createJsonRpcRequest(req.method, encodeBroadcastTxParams(req.params));
   }
 
   public static encodeCommit(req: requests.CommitRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeHeightParam(req.params));
+    return createJsonRpcRequest(req.method, encodeHeightParam(req.params));
   }
 
   public static encodeGenesis(req: requests.GenesisRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method);
+    return createJsonRpcRequest(req.method);
   }
 
   public static encodeHealth(req: requests.HealthRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method);
+    return createJsonRpcRequest(req.method);
   }
 
   public static encodeStatus(req: requests.StatusRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method);
+    return createJsonRpcRequest(req.method);
   }
 
   public static encodeSubscribe(req: requests.SubscribeRequest): JsonRpcRequest {
     const eventTag = { key: "tm.event", value: req.query.type };
     const query = requests.buildQuery({ tags: [eventTag], raw: req.query.raw });
-    return jsonRpcWith("subscribe", { query: query });
+    return createJsonRpcRequest("subscribe", { query: query });
   }
 
   public static encodeTx(req: requests.TxRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeTxParams(req.params));
+    return createJsonRpcRequest(req.method, encodeTxParams(req.params));
   }
 
   // TODO: encode params for query string???
   public static encodeTxSearch(req: requests.TxSearchRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeTxSearchParams(req.params));
+    return createJsonRpcRequest(req.method, encodeTxSearchParams(req.params));
   }
 
   public static encodeValidators(req: requests.ValidatorsRequest): JsonRpcRequest {
-    return jsonRpcWith(req.method, encodeHeightParam(req.params));
+    return createJsonRpcRequest(req.method, encodeHeightParam(req.params));
   }
 }