+++ packages/iov-tendermint-rpc/types/rpcclients/index.d.ts
@@ -1,3 +1,3 @@
-export { instanceOfRpcStreamingClient, RpcClient, RpcStreamingClient } from "./rpcclient";
+export { instanceOfRpcStreamingClient, RpcClient, RpcStreamingClient, SubscriptionEvent } from "./rpcclient";
 export { HttpClient } from "./httpclient";
 export { WebsocketClient } from "./websocketclient";