+++ packages/iov-tendermint-rpc/types/rpcclients/websocketclient.d.ts
@@ -1,6 +1,6 @@
 import { Stream } from "xstream";
-import { JsonRpcEvent, JsonRpcRequest, JsonRpcResponse, JsonRpcSuccess } from "../jsonrpc";
-import { RpcStreamingClient } from "./rpcclient";
+import { JsonRpcId, JsonRpcRequest, JsonRpcResponse, JsonRpcSuccessResponse } from "@iov/jsonrpc";
+import { RpcStreamingClient, SubscriptionEvent } from "./rpcclient";
 export declare class WebsocketClient implements RpcStreamingClient {
     private readonly url;
     private readonly socket;
@@ -8,13 +8,13 @@ export declare class WebsocketClient implements RpcStreamingClient {
     private readonly jsonRpcResponseStream;
     private readonly subscriptionStreams;
     constructor(baseUrl?: string, onError?: (err: any) => void);
-    execute(request: JsonRpcRequest): Promise<JsonRpcSuccess>;
-    listen(request: JsonRpcRequest): Stream<JsonRpcEvent>;
+    execute(request: JsonRpcRequest): Promise<JsonRpcSuccessResponse>;
+    listen(request: JsonRpcRequest): Stream<SubscriptionEvent>;
     /**
      * Resolves as soon as websocket is connected. execute() queues requests automatically,
      * so this should be required for testing purposes only.
      */
     connected(): Promise<void>;
     disconnect(): void;
-    protected responseForRequestId(id: string): Promise<JsonRpcResponse>;
+    protected responseForRequestId(id: JsonRpcId): Promise<JsonRpcResponse>;
 }