+++ CHANGELOG.md
@@ -31,6 +31,10 @@ Breaking changes
   `TransactionId | null`.
 - @iov/ethereum: Switch to new `SwapId` type instead of `SwapIdBytes` where
   appropriate.
+- @iov/jsonrpc: The `id` field of `JsonRpcRequest`, `JsonRpcSuccessResponse` and
+  `JsonRpcErrorResponse` is now of type `JsonRpcId`, which can be a string or a
+  number.
+- @iov/jsonrpc: Remove `parseJsonRpcResponse` and `parseJsonRpcError`.
 
 ## 0.14.1